add HA controller details to ZitiCtxEvent

+++ library/bind.c
@@ -432,7 +432,7 @@ static void bind_reply_cb(void *ctx, message *msg, int code) {
     if (code == ZITI_OK && msg->header.content == ContentTypeStateConnected) {
         CONN_LOG(TRACE, "received msg ct[%X] code[%d]", msg->header.content, code);
         CONN_LOG(DEBUG, "bound successfully over ch[%s]", b->ch->url);
-        ziti_channel_add_receiver(b->ch, (int)conn->conn_id, b,
+        ziti_channel_add_receiver(b->ch, conn->conn_id, b,
                                   (void (*)(void *, message *, int)) on_message);
         b->bound = true;
     } else {
